// src/app/services/model/credentials.ts
/**
 * VAmPI
 * OpenAPI v3 specs for VAmPI
 *
 * OpenAPI spec version: 0.1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface Credential { 
    username?: string;
    password?: string;
}